# hypot scaling constant folding