The scaling constant in `hypot` wasn't being constant folded. Partly fixes #36353.