@simonbyrne 2020-06-19
Before:
```
julia> @benchmark hypot(1.0,2.0)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     11.176 ns (0.00% GC)
  median time:      11.243 ns (0.00% GC)
  mean time:        12.354 ns (0.00% GC)
  maximum time:     113.271 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     998
```
After:
```
julia> @benchmark xhypot(1.0,2.0)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     7.083 ns (0.00% GC)
  median time:      7.304 ns (0.00% GC)
  mean time:        7.946 ns (0.00% GC)
  maximum time:     73.317 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     999
```