@KlausC 2020-06-23
Do you consider these ideas of @kimikage ? https://github.com/JuliaLang/julia/issues/36353#issuecomment-646825768